refactor: PostsMangerPage 컴포넌트 widgets로 분할
